Move GPU post-processing separately to cuda_postprocess.cu
