+++ yolov8/include/preprocess.h
@@ -6,12 +6,6 @@
 #include <map>
 
 
-struct AffineMatrix {
-    float value[6];
-};
-
-const int bbox_element = sizeof(AffineMatrix) / sizeof(float)+1;      // left, top, right, bottom, confidence, class, keepflag
-
 void cuda_preprocess_init(int max_image_size);
 
 void cuda_preprocess_destroy();
@@ -20,6 +14,3 @@ void cuda_preprocess(uint8_t *src, int src_width, int src_height, float *dst, in
 
 void cuda_batch_preprocess(std::vector<cv::Mat> &img_batch, float *dst, int dst_width, int dst_height, cudaStream_t stream);
 
-void cuda_decode(float* predict, int num_bboxes, float confidence_threshold,float* parray,int max_objects, cudaStream_t stream);
-
-void cuda_nms(float* parray, float nms_threshold, int max_objects, cudaStream_t stream);
